@@ -4,10 +4,12 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
 import android.view.View;
 import android.widget.Button;
+import android.widget.EditText;
 import android.widget.Toast;
+
+import androidx.annotation.Nullable;
 
 import com.tokenpocket.opensdk.base.TPListener;
 import com.tokenpocket.opensdk.base.TPManager;
@@ -25,6 +27,13 @@ import tokenpocket.pro.sdk_demo.R;
 
 public class EthTransferActivity extends Activity implements View.OnClickListener {
 
+    private EditText etFrom;
+    private EditText etTo;
+    private EditText etAmount;
+    private EditText etContract;
+    private EditText etSymbol;
+    private EditText etDecimal;
+    private EditText etChainId;
     private Button btnTransfer;
 
     @Override
@@ -32,7 +41,14 @@ public class EthTransferActivity extends Activity implements View.OnClickListene
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_eth_transfer);
 
+        etFrom = findViewById(R.id.et_from);
+        etTo = findViewById(R.id.et_to);
+        etAmount = findViewById(R.id.et_amount);
+        etContract = findViewById(R.id.et_contract);
+        etSymbol = findViewById(R.id.et_symbol);
+        etDecimal = findViewById(R.id.et_decimal);
         btnTransfer = findViewById(R.id.btn_transfer);
+        etChainId = findViewById(R.id.et_chain_id);
 
         btnTransfer.setOnClickListener(this);
     }
@@ -46,16 +62,16 @@ public class EthTransferActivity extends Activity implements View.OnClickListene
         }
     }
 
-
     /**
      * 这里以BSC上转账实例，说明各个参数
      */
     private void transfer() {
+        String chainId = etChainId.getText().toString();
         Transfer transfer = new Transfer();
         //标识链
         List<Blockchain> blockchains = new ArrayList<>();
         //evm系列，第一个参数是ethereum,第二个参数是网络的id，这里56是BSC网络链上id
-        blockchains.add(new Blockchain("ethereum", "56"));
+        blockchains.add(new Blockchain("ethereum", chainId));
         transfer.setBlockchains(blockchains);
 
         transfer.setProtocol("TokenPocket");
@@ -64,21 +80,19 @@ public class EthTransferActivity extends Activity implements View.OnClickListene
         transfer.setDappIcon("https://eosknights.io/img/icon.png");
         //开发者自己定义的业务Id,用来标识这次操作
         transfer.setActionId("web-db4c5466-1a03-438c-90c9-2172e8becea5");
-        //data，如果是转原生代币，可以添加上链数据
-        transfer.setMemo("0xe595a6");
-        transfer.setAction("transfer");
-        //发送者
-        transfer.setFrom("0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5");
-        //接受者
-        transfer.setTo("0x32ff06198da462f1c519d30f4d328b3fef295d19");
-        //代币合约地址，如果是转ETH，可以不设置这个参数
-        transfer.setContract("0xdAC17F958D2ee523a2206206994597C13D831ec7");
-        //转账数量，比如这里demo是转0.01个USDT，就是传入0.01
-        transfer.setAmount(0.01);
-        //必须设置
-        transfer.setDecimal(18);
-        transfer.setSymbol("USDT");
-        transfer.setDesc("仅供ui展示，不上链");
+
+        transfer.setFrom(etFrom.getText().toString());
+        transfer.setTo(etTo.getText().toString());
+        transfer.setContract(etContract.getText().toString());
+        transfer.setAmount(getAmount());
+        transfer.setSymbol(etSymbol.getText().toString());
+        transfer.setDecimal(Integer.valueOf(etDecimal.getText().toString()));
+
+        //代币合约地址，如果是转原生币，不需要设置这个参数
+        //        transfer.setContract("0x55d398326f99059ff775485246999027b3197955");
+        //如果是转原生币，可以添加上链数据
+        //        transfer.setMemo("0xe595a6");
+
         //开发者服务端提供的接受调用登录结果的接口，如果设置该参数，钱包操作完成后，会将结果通过post application json方式将结果回调给callbackurl
         transfer.setCallbackUrl("http://115.205.0.178:9011/taaBizApi/taaInitData");
         TPManager.getInstance().transfer(this, transfer, new TPListener() {
@@ -102,6 +116,14 @@ public class EthTransferActivity extends Activity implements View.OnClickListene
         });
     }
 
+    private double getAmount() {
+        try {
+            return Double.parseDouble(etAmount.getText().toString());
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return 0;
+    }
 
     public static void start(Context context) {
         Intent intent = new Intent(context, EthTransferActivity.class);
