@@ -2,15 +2,15 @@ package tokenpocket.pro.sdk_demo.minwallet;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
 
 
+import androidx.annotation.Nullable;
+
 import com.tokenpocket.opensdk.base.TPListener;
 import com.tokenpocket.opensdk.base.TPManager;
-//import com.tokenpocket.opensdk.net.custom.Json;
 import com.tokenpocket.opensdk.simple.model.Transaction;
 
 import tokenpocket.pro.sdk_demo.R;
