@@ -3,11 +3,12 @@ package tokenpocket.pro.sdk_demo.tron;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
-import android.widget.EditText;
 import android.widget.Toast;
+
+import androidx.annotation.Nullable;
 
 import com.tokenpocket.opensdk.base.TPListener;
 import com.tokenpocket.opensdk.base.TPManager;
@@ -176,30 +177,31 @@ public class TronDemoActivity extends Activity implements View.OnClickListener {
          * }
          */
         transaction.setTxData("{\n" +
-                "\t\"visible\": false,\n" +
-                "\t\"txID\": \"7d4a8868a806aaf2a3f5bf23f51667d83bf67bcaf17d3e176a4ce347a2450649" +
-                "\",\n" +
-                "\t\"raw_data\": {\n" +
-                "\t\t\"contract\": [{\n" +
-                "\t\t\t\"parameter\": {\n" +
-                "\t\t\t\t\"value\": {\n" +
-                "\t\t\t\t\t\"frozen_duration\": 3,\n" +
-                "\t\t\t\t\t\"frozen_balance\": 4000000,\n" +
-                "\t\t\t\t\t\"owner_address\": \"41593a115d4e04249a7e4c02a2150b596081b37f9f\"\n" +
+                "\t\t\"visible\": false,\n" +
+                "\t\t\"txID\": \"7111456f4e7e86360790a8d10e8e3e78e7f882cfd6f190805a78a1e5da335705\",\n" +
+                "\t\t\"raw_data\": {\n" +
+                "\t\t\t\"contract\": [{\n" +
+                "\t\t\t\t\"parameter\": {\n" +
+                "\t\t\t\t\t\"value\": {\n" +
+                "\t\t\t\t\t\t\"data\": " +
+                "\"c10bea5c000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000001c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003487b63d30b5b2c87fb7ffa8bcfade38eaac1abe00000000000000000000000087d5ef4c42a78f85dd1ca399fa0c99f9ab0875d9000000000000000000000000f90a4115ca0859c0db8415d73b3a22626506cbbe000000000000000000000000000000000000000000000000000000000098968000000000000000000000000000000000000000000000000000000000000ad4a400000000000000000000000000000000000000000000000000000000000001400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000891cdb91d149f23b1a45d9c5ca78a88d0cb44c180000000000000000000000000000000000000000000000000000000000000003776562000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000002e000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f90a4115ca0859c0db8415d73b3a22626506cbbe00000000000000000000000000000000000000000000000000000000000000e00000000000000000000000000000000000000000000000000000000000000120000000000000000000000000000000000000000000000000000000000000016000000000000000000000000000000000000000000000000000000000000001a000000000000000000000000000000000000000000000000000000000000001e000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000982150000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000ec5e8661f8855b1724d4f668f11ab722b4b3b74d00000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000064ce22f85800000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000064df3808000000000000000000000000f90a4115ca0859c0db8415d73b3a22626506cbbe00000000000000000000000000000000000000000000000000000000\",\n" +
+                "\t\t\t\t\t\t\"owner_address\": \"41f90a4115ca0859c0db8415d73b3a22626506cbbe\",\n" +
+                "\t\t\t\t\t\t\"contract_address\": \"416c37365d9f04aa507c019e91849a6c61584c9ef2\",\n" +
+                "\t\t\t\t\t\t\"call_value\": 10000000\n" +
+                "\t\t\t\t\t},\n" +
+                "\t\t\t\t\t\"type_url\": \"type.googleapis.com/protocol.TriggerSmartContract\"\n" +
                 "\t\t\t\t},\n" +
-                "\t\t\t\t\"type_url\": \"type.googleapis.com\\/protocol.FreezeBalanceContract\"\n" +
-                "\t\t\t},\n" +
-                "\t\t\t\"type\": \"FreezeBalanceContract\"\n" +
-                "\t\t}],\n" +
-                "\t\t\"ref_block_bytes\": \"ebec\",\n" +
-                "\t\t\"ref_block_hash\": \"83e6e13e48ce9f6c\",\n" +
-                "\t\t\"expiration\": 1648893501000,\n" +
-                "\t\t\"timestamp\": 1648893443731\n" +
-                "\t},\n" +
-                "\t\"raw_data_hex\": " +
-                "\"0a02ebec220883e6e13e48ce9f6c40c8bcd8cdfe2f5a58080b12540a32747970652e676f6f676c65617069732e636f6d2f70726f746f636f6c2e467265657a6542616c616e6365436f6e7472616374121e0a1541593a115d4e04249a7e4c02a2150b596081b37f9f108092f40118037093fdd4cdfe2f\",\n" +
-                "\t\"userTronHeader\": true\n" +
-                "}");
+                "\t\t\t\t\"type\": \"TriggerSmartContract\"\n" +
+                "\t\t\t}],\n" +
+                "\t\t\t\"ref_block_bytes\": \"9571\",\n" +
+                "\t\t\t\"ref_block_hash\": \"cde1f067d8565a7e\",\n" +
+                "\t\t\t\"expiration\": 1692349329000,\n" +
+                "\t\t\t\"fee_limit\": 150000000,\n" +
+                "\t\t\t\"timestamp\": 1692349272590\n" +
+                "\t\t},\n" +
+                "\t\t\"raw_data_hex\": " +
+                "\"0a0295712208cde1f067d8565a7e40e8dc85bfa0315a950b081f12900b0a31747970652e676f6f676c65617069732e636f6d2f70726f746f636f6c2e54726967676572536d617274436f6e747261637412da0a0a1541f90a4115ca0859c0db8415d73b3a22626506cbbe1215416c37365d9f04aa507c019e91849a6c61584c9ef21880ade20422a40ac10bea5c000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000001c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003487b63d30b5b2c87fb7ffa8bcfade38eaac1abe00000000000000000000000087d5ef4c42a78f85dd1ca399fa0c99f9ab0875d9000000000000000000000000f90a4115ca0859c0db8415d73b3a22626506cbbe000000000000000000000000000000000000000000000000000000000098968000000000000000000000000000000000000000000000000000000000000ad4a400000000000000000000000000000000000000000000000000000000000001400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000891cdb91d149f23b1a45d9c5ca78a88d0cb44c180000000000000000000000000000000000000000000000000000000000000003776562000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000002e000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f90a4115ca0859c0db8415d73b3a22626506cbbe00000000000000000000000000000000000000000000000000000000000000e00000000000000000000000000000000000000000000000000000000000000120000000000000000000000000000000000000000000000000000000000000016000000000000000000000000000000000000000000000000000000000000001a000000000000000000000000000000000000000000000000000000000000001e000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000982150000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000ec5e8661f8855b1724d4f668f11ab722b4b3b74d00000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000064ce22f85800000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000064df3808000000000000000000000000f90a4115ca0859c0db8415d73b3a22626506cbbe00000000000000000000000000000000000000000000000000000000708ea482bfa031900180a3c347\"\n" +
+                "\t}");
         //开发者服务端提供的接受调用登录结果的接口，如果设置该参数，钱包操作完成后，会将结果通过post application json方式将结果回调给callbackurl
         transaction.setCallbackUrl("http://115.205.0.178:9011/taaBizApi/taaInitData");
         TPManager.getInstance().pushTransaction(this, transaction, new TPListener() {
@@ -207,6 +209,7 @@ public class TronDemoActivity extends Activity implements View.OnClickListener {
             public void onSuccess(String s) {
                 //转账成功后，会返回相应的交易hash，注意，钱包只是把交易push出去，并不能保证最后交易结果，开发者需要根据hash自行确定链上结果
                 Toast.makeText(TronDemoActivity.this, s, Toast.LENGTH_LONG).show();
+                Log.e("pushTransaction", s);
             }
 
             @Override
